JS Long (#328)

+++ doordeck-sdk/src/jsMain/kotlin/com/doordeck/multiplatform/sdk/model/responses/FusionResponses.js.kt
@@ -143,7 +143,7 @@ internal fun BasicLockController.toLockControllerResponse(): LockControllerRespo
 internal fun BasicAlpetaController.toAlpetaController(): FusionOperations.AlpetaController = FusionOperations.AlpetaController(
     username = username,
     password = password,
-    doorId = doorId.toInt(),
+    doorId = doorId,
     baseUrl = baseUrl
 )
